Rename `MenuPage` to `Menu`
